Add files via upload
The Problem
Statement
As we see nowadays organic farming is widely practicing by the
people. It involves storing all the biodegradable matter such as
vegetables waste and rotten leaves and allowing microorganisms
to grow. The problem of making the compost is it need proper
maintaince and great observation to make a good organic
compost. The challenge with this approach is the care needed to
maintain the composting environment. This involves aerating the
compost, adding materials to maintain heat/pH, and watering the
compost as needed. This is where the Smart Compost System can
help.

solution:
This problem can be resolved by making a bin which stores
the waste and make quality compost. The smart composting
system monitors and manages your compost - adding air
and water when needed. It tracks when your compost is
ready and alerts you when a specific action needs to be
taken like stir the compost, cover the compost. It's simple
way of generating compost at home itself. If all you did was
dump kitchen scraps and leaves into a pile, then you're
guaranteed to get compost.